Adapt generators with Controller.GetDirectInputMapping
